Dont overwrite config files that already exist

+++ app/imagestart.sh
@@ -53,11 +53,25 @@ fi
 mkdir -p ${HOME}/.vnc
 install -m 0755 /usr/share/ubuntu-desktop/vnc/xstartup ${HOME}/.vnc/xstartup
 
-# Openbox configuration
+# Openbox/FBPanel configuration
 if [ "${WINDOWMANAGER}" == "openbox" ]; then
-  mkdir -p ${HOME}/.config/{openbox,fbpanel}
-  install -m 0755 /usr/share/ubuntu-desktop/openbox/autostart ${HOME}/.config/openbox/autostart
-  install -m 0644 /usr/share/ubuntu-desktop/openbox/menu.xml ${HOME}/.config/openbox/menu.xml
-  install -m 0644 /usr/share/ubuntu-desktop/openbox/rc.xml ${HOME}/.config/openbox/rc.xml
-  install -m 0644 /usr/share/ubuntu-desktop/fbpanel/default ${HOME}/.config/fbpanel/default
+  if [ ! -d ${HOME}/.config/openbox ]; then
+    mkdir -p ${HOME}/.config/openbox
+  fi
+  if [ ! -f ${HOME}/.config/openbox/autostart ]; then
+    install -m 0755 /usr/share/ubuntu-desktop/openbox/autostart ${HOME}/.config/openbox/autostart
+  fi
+  if [ ! -f ${HOME}/.config/openbox/menu.xml ]; then
+    install -m 0644 /usr/share/ubuntu-desktop/openbox/menu.xml ${HOME}/.config/openbox/menu.xml
+  fi
+  if [ ! -f ${HOME}/.config/openbox/rc.xml ]; then
+    install -m 0644 /usr/share/ubuntu-desktop/openbox/rc.xml ${HOME}/.config/openbox/rc.xml
+  fi
+
+  if [ ! -d ${HOME}/.config/fbpanel ]; then
+    mkdir -p ${HOME}/.config/fbpanel
+  fi
+  if [ ! -f ${HOME}/.config/fbpanel/default ]; then
+    install -m 0644 /usr/share/ubuntu-desktop/fbpanel/default ${HOME}/.config/fbpanel/default
+  fi
 fi
